new_map 2.9

+++ r_c.py
@@ -1,8 +1,8 @@
 import pygame
 from parameters import *
 from numba import njit
-from map import txt_map, H_WORLD, W_WORLD
-
+#from map import txt_map, H_WORLD, W_WORLD
+from new_map import txt_map, H_WORLD, W_WORLD
 
 @njit(fastmath=True)
 def mapping(a, b):
@@ -47,7 +47,7 @@ def ray_casting(pos_gamer, angle_gamer, txt_map):
         offset = int(offset) % CELL
         depth *= math.cos(angle_gamer - view_angle)
         depth = max(depth, 0.00001)
-        hight = (PROJ_C / depth)
+        hight = int(PROJ_C / depth)
 
         walls.append((depth, offset, hight, texture))
         view_angle += DELTA_ANGLE
@@ -65,11 +65,11 @@ def walls_with_ray_cast(gamer, textures):
             texture_height = T_H / coeff
             wall_c = textures[texture].subsurface(offset * T_SCALE, H_T_H - texture_height // 2,
                                                   T_SCALE, texture_height)
-            wall_c = pygame.transform.scale(wall_c, (int(SCALE), int(HEIGHT)))
+            wall_c = pygame.transform.scale(wall_c, (SCALE, HEIGHT))
             walls_pos = (ray * SCALE, 0)
         else:
             wall_c = textures[texture].subsurface(offset * T_SCALE, 0, T_SCALE, T_H)
-            wall_c = pygame.transform.scale(wall_c, (int(SCALE), int(hight)))
+            wall_c = pygame.transform.scale(wall_c, (SCALE, hight))
             walls_pos = (ray * SCALE, H_HEIGHT - hight // 2)
         walls.append((depth, wall_c, walls_pos))
     return walls, wall_shot
